Added a new program, rangefinder
This program is made for the MaxBotics
Ultrasonic sensor. In the main robot program is how to initialize it, as well as the code to make it work. Also started to work on the network tables
The rangefinder program is rewritten C++ code which can be found at https://www.chiefdelphi.com/forums/showthread.php?t=82409